requirements.txt
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.5
urllib3==2.2.1
Werkzeug==3.0.2
